Rewrite dMAC as the gateway's only for local and remote Pods
It is a workaround for a Windows OVS issue which marks the reply packet
as "new" in conntrack, and mark it as a connection initiated from
antrea-gw0 (including the Node or the external). The issue results the
reply packet from antrea-gw0 is output to antrea-gw0 and finally dropped
by OVS datapath because the input and output port number are the same.

The workaround is only rewrite the dst MAC as antrea-gw0's on the reply
packet which enters OVS from the local Pods or the tunnel port even if
the connection is marked as initiated from antrea-gw0.

Signed-off-by: wenyingd <wenyingd@vmware.com>