Make StartTime a pointer to metav1.Time
To avoid issues with OpenAPI validation in K8s versions prior to v1.20.
See https://github.com/kubernetes/kubernetes/issues/86811
The recommendation to avoid issues with metav1.Time being serialized as
null when the field is unset seems to be to make it a pointer. A nil
pointer is them omitted during serialization. As far as I can tell, it
should also be possible to stick with a metav1.Time and make the
property "nullable" in the OpenAPI schema, but I am not sure whether
this can create other issues.

Signed-off-by: Antonin Bas <abas@vmware.com>