Add containerName param to probe func
This PR has two changes:

1. Adds a parameter `containerName` to `probe` func.

Previously we use "c[port]" as containerName in `probe` to call
`runCommandFromPod`. But in this case, `probe` func couldn't work
with the client created by `createAgnhostPodOnNode` or functions that
create Pod with different container names. We hope `probe` func could
be more generic, so I add `containerName` parameter. Let the caller
to choose the containerName.

2. Changes the way that how we decides the connectivity.

Our framework will do probe 3 times to get the connectivity. Previous
logic is:
  - If one of them connected, then it's connected.
  - If no one connected and one of them is timeout, then it's
    dropped.
  - All other situation return rejected.

This could be a problem. For example: `testRejectServiceTraffic`
creates an agnhost Pod as client and use `probe` func to do the
probe. But agnhost Pod doesn't have "c[port]" container. So this
probe falls into the third case. It will always return rejected which
is also what we expected! Also, sometimes, due to the pressure on our
testbed or other test-side timing issues, it will return timeout. We
shouldn't take this as a `dropped`.

Now change the the logic to:
  - If no one connected and at least one got rejected, then return
    rejected.
  - If no one rejected and at least one connected, then return
    connected.
  - If all probes timeout, then return dropped.
  - All other sithation return error.

Signed-off-by: wgrayson <wgrayson@vmware.com>