Add startTime to the Traceflow Status (#2952)
* Add startTime to the Traceflow Status

The startTime is used to determine if a Traceflow has timed out and
should be reported as failed.

Until now we were relying on the CreationTimestamp set by the
kube-apiserver. However, by chance, I was testing Traceflow on a cluster
with a clock skew between the control plane Node (running the
kube-apiserver) and the worker Node running the
antrea-controller. Because of the clock skew, each Traceflow request was
tagged as failed as soon as I created it (with reason timeout). It took
me a while to figure out the reason.

By introducing the startTime field (set & used by the
antrea-controller), we avoid the possibility of such issues. If
startTime is not available for any reason, we fall back to the
CreationTimestamp.

This API change is backwards-compatible.

Signed-off-by: Antonin Bas <abas@vmware.com>

* Make StartTime a pointer to metav1.Time

To avoid issues with OpenAPI validation in K8s versions prior to v1.20.
See https://github.com/kubernetes/kubernetes/issues/86811
The recommendation to avoid issues with metav1.Time being serialized as
null when the field is unset seems to be to make it a pointer. A nil
pointer is then omitted during serialization. As far as I can tell, it
should also be possible to stick with a metav1.Time and make the
property "nullable" in the OpenAPI schema, but I am not sure whether
this can create other issues.

Signed-off-by: Antonin Bas <abas@vmware.com>