common area and clusterset reconciler
ClusterSet reconciler performs reconcilation of cluster set resource

MCS Controller currently needs to be run in a leader/member mode.
Because the operations performed on leader/member are different
this is taken as an argument to the controller command. In a future
patch this will be implemented as two separate commands to avoid
multiple conditional checks.

Validations:
-----------
Following validations are performed by the ClusterSet Reconciler
1. There must be a valid ClusterClaim with ClusterID and ClusterSetID
2. The configured ClusterID is defined as either a member/leader in the
ClusterSet
3. Only one ClusterSet configuration is allowed

On the member cluster:
---------------------
ClusterSet reconciler will initialize RemoteClusterManager to
manage connections to all Leaders which will be represented as a
RemoteCluster. Once a connection is established two things occur:
1. A MemberClusterAnnounce resource is written to all Leader API
Servers
2. A leader election is performed to choose an elected leader
among all configured leaders.
An elected leader is the leader cluster from where resources
will be imported for the resource exchange pipeline.

Common Area refers to the namespace within a leader cluster
where the leader controller is running and where resources are
written and read from.

On the leader cluster:
---------------------
ClusterSet reconciler mainly performs the validation.

Status:
------
Status of the ClusterSet resource is not implemented in this
patch and will come in a future patch.

Signed-off-by: Aravinda Kidambi <akidambi@vmware.com>