Manifest generation logic
There are 3 manifests that are generated:
1. Leader global - this contains all the global resources that needs
to be applied once on the leader cluster.
2. Leader namespaced - the leader MCS controller runs within the
context of a namespace. Namespaced yaml must be generated for the
required namespace and applied on the leader cluster. The provided
yaml uses `changeme` as the namespace for all resources and must be
changed to the correct namespace required. Or a new yaml can also be
generated using
$ antrea/multicluster/hack/generate-manifest.sh -l <namespace>
3. Member - this is the manifest that needs to be applied on the
member cluster.

Key differences in Member and Leader namespaced yamls:
Controller:
----------
A member runs only one MCS controller which reconciles resources
across all namespaces in the cluster.
A leader runs one MCS controller in a given namspace which
reconciles resources only within that namespace.

RBAC is configured to provide entire cluster access to the member
controller and for the given namespace for the leader controller

Webhooks:
---------
Webhooks are configured globally in the member deployment.
Webhooks are configured within a namespace in the leader deployment.
So in leader cluster which has multiple MCS deployments, there will
be multiple webhooks each redirecting requests within a given
namespace to the corresponding controller. This was a concious choice
to avoid a hierarchical controller. Also webhooks are performing
namespace specific validations so they cannot be handled by a
global service.

A cluster that is both a leader and member:
-----------------------------------------
A cluster can be both a leader and member, in this case, there
will be two controllers running - one member controller and second
leader controller within a namespace.
TODO: We need special handling so member controller does not
reconcile resources within leader namespace even though it has
cluster-wide access.

Signed-off-by: Aravinda Kidambi <akidambi@vmware.com>