[Windows] Use Windows management virtual network adapter for OVS bridge
Rename the vnic created by Windows host when creating the HNS Network as
the OVS bridge name, and add OVS internal port for the bridge port. Then
OVS uses the vnic created by Windows host directly, and IP/MAC/route
migrations is not needed. With this workflow, the time of losing IP on
the Windows host is reduced.

Signed-off-by: wenying <wenyingd@vmware.com>