Fix NetworkPolicy resources dump for Agent's supportbundle
There was an obvious bug where all NetworkPolicy resources
(NetworkPolicies, AddressGroups, AppliedToGroups) were dumped to the
same file, and the file itself ("agentinfo") was subsequently
overwritten with a dump of the appropriate AgentInfo CRD as expected.

This is fixed by dumping these resources to their respective appropriate
files: networkpolicies, addressgroups, appliedtogroups.

In addition, we:
* switch from JSON to YAML for these resources (along with the AgentInfo
  CRD) for consistency with the files included in the Controller's
  supportbundle.
* define a custom MarshalYAML function for the IPAddress type. IPAddress
  is typedef'ed to a byte slice, which means that by default it is
  marshalled as an array of individual byte values, making the files
  difficult to read.

Fixes #3082

Signed-off-by: Antonin Bas <abas@vmware.com>