Fix status report of Antrea-native policies (#3074)
For an Antrea-native policy with multiple rules that have different
AppliedTo, some rules may not be effective on a Node when the policy
applies to the Node. In this case, antrea-agent didn't make any actual
dataplane change and didn't report new status to antrea-controller if
the only changed part of a policy event is about the ineffective rules,
even though the generation of the policy may have changed. This leads
to a policy staying in realizing state because of the missing status
report.

This patch fixes it by triggering status report at least once as long as
a policy is actually updated. In addition, it removes some useless
return value and changes the order of event logs to better reflect the
cause of policy reconciliation.

Signed-off-by: Quan Tian <qtian@vmware.com>