Common area and ClusterSet reconciler (#3066)
MCS CommonArea and ClusterSetReconciler implementation

MCS is a feature that allows resource and policy configuration
that span across multiple k8 clusters. A ClusterSet defines such
a group of k8 clusters which requires resource exchange.

A ClusterSet needs to have one or more clusters chosen as a
Leader which facilitates the resource exchange so that we dont
need a full-mesh between all clusters in the ClusterSet. Leader
can be dedicated which means they dont participate in the
resource exchange themselves. Or they call be both a leader
and member, so it facilitates and also takes part in the resource
exchange.

Members of the ClusterSet write certain resources into leader
cluster’s common area and all members pull resources from
that common area. Common area refers to the namespace
within a leader cluster where the leader controller is running
and where resources are written and read from.

MCS Controller needs to be run in a leader/member mode.
Because the operations performed on leader/member are different
two sub-commands (leader, member) are provided to specify
which mode the controller is running in.

Validations for ClusterSet configuration
----------------------------------------
Following validations are performed by the ClusterSetReconciler
1. There must be a valid ClusterClaim with ClusterID and ClusterSetID
2. The configured ClusterID is defined as either a member/leader in the
ClusterSet
3. Only one ClusterSet configuration is allowed

MemberClusterSetReconciler
------------------------------
MemberClusterSetReconciler runs only in member cluster
deployment. It will initialize RemoteCommonAreaManager
to manage connections to all Leader’s common area which
will be represented as RemoteCommonArea. Once a connection
is established two things occur:
1. A MemberClusterAnnounce resource is written to all 
    RemoteCommonArea periodically.
2. A leader election is performed to choose an elected leader
    among all configured leaders.
An elected leader is the leader cluster from where resources
will be imported for the resource exchange pipeline.

LeaderClusterSetReconciler
-----------------------------
LeaderClusterSetReconciler runs only in leader cluster
deployment and mainly performs the validations.

ClusterSet status:
------------------
Status of the ClusterSet resource is not implemented in this
patch and will come in a future patch.

MCS Manifests
---------------
There are 3 manifests that are generated:
1. Leader global - this contains all the global resources that needs
to be applied once on the leader cluster.
2. Leader namespaced - the leader MCS controller runs within the
context of a namespace. Namespaced yaml must be generated for the
required Namespace and applied on the leader cluster. The provided
yaml uses `changeme` as the Namespace for all resources and must be
changed to the correct Namespace required. Or a new yaml can also be
generated using
 $ antrea/multicluster/hack/generate-manifest.sh -l <namespace>
3. Member - this is the manifest that needs to be applied on the
member cluster.

Key differences in Member and Leader Namespaced yamls:
Controller:
----------
A member runs only one MCS controller which reconciles resources
across all namespaces in the cluster.
A leader runs one MCS controller in a given Namespace which
reconciles resources only within that Namespace.
RBAC is configured to provide entire cluster access to the member
controller and for the given Namespace for the leader controller.

Webhooks:
-----------
Webhooks are configured globally in the member deployment.
Webhooks are configured within a Namespace in the leader Deployment.
So in leader cluster which has multiple MCS Deployments, there will
be multiple webhooks each redirecting requests within a given
Namespace to the corresponding controller. This was a conscious choice
to avoid a hierarchical controller. Also webhooks will be performing
Namespace specific validations so they cannot be handled by a
global service.

A cluster that is both a leader and member:
-------------------------------------------
A cluster can be both a leader and member, in this case, there
will be two controllers running - one member controller and second
leader controller within a Namespace.
TODO: We need special handling so member controller does not
reconcile resources within leader Namespace even though it has
cluster-wide access.

Unit testing:
=========
RemoteCommonArea Tests
---------------------------
Test that MemberClusterAnnounce is written and connectivity is set
to true.

TODO: Find out if fakeClient can error writes so we can test
disconnection case

NOTE: used mockgen to generate mocks for Manager from controller-runtime
library and for RemoteCommonAreaManager. TODO: Add mockgen to Makefile
so it always gets generated automatically.

RemoteCommonAreaManager and LeaderElector tests
-------------------------------------------------------
1. Test adding a new leader cluster
2. Test leader election
3. Test leader election when elected leader becomes disconnected

TODO: Test removing a leader cluster

NOTE: Used mockgen to generate a mock for RemoteCluster so that
we can verify StartMonitoring is invoked. TODO: Add mockgen to Makefile
so it always gets generated automatically.

Signed-off-by: Aravinda Kidambi <akidambi@vmware.com>