Fix IPsec for IPv6 overlays
When using IPv6, the IPsec configuration (ipsec.conf) generated by
ovs-monitor-ipsec for strongSwan is currently not correct. A patch has
been submitted upstream, but until it is accepted and merged, we apply a
temporary version of the patch.

This was tested for a VXLAN overlay in an IPv6-only cluster.

Fixes #3151

Signed-off-by: Antonin Bas <abas@vmware.com>