Add env parameter to support noEncap without AntreaProxy (#3116)
When using NoEncap traffic mode without AntreaProxy, Pod-to-Service
traffic is handled by kube-proxy (iptables/ipvs) in the root netns.
If the Endpoint is not local the DNATed traffic will be output to
the physical network directly without going back to OVS for Egress
NetworkPolicy enforcement, which breaks basic security functionality.
Therefore, we usually do not allow the NoEncap traffic mode without
AntreaProxy. But one can bypass this check and force this feature
combination to be allowed, by defining the ALLOW_NO_ENCAP_WITHOUT_ANTREA_PROXY
environment variable and setting it to true. This may lead to better
performance when using NoEncap if Egress NetworkPolicy enforcement is
not required.

Signed-off-by: Wenze Gao <wenze.gao@transwarp.io>
Signed-off-by: Wu zhengdong <zhengdong.wu@transwarp.io>

Co-authored-by: Wenze Gao <wenze.gao@transwarp.io>