Make LoadBalancer IP proxying configurable for AntreaProxy (#3130)
A new proxyLoadBalancerIPs configuration parameter is added for
AntreaProxy. When set to false, AntreaProxy no longer load-balances
traffic destined to the External IPs of LoadBalancer Services. This is
useful when the external LoadBalancer provides additional capabilities
(e.g. TLS termination) and it is desirable for Pod-to-ExternalIP traffic
to be sent to the external LoadBalancer instead of being load-balanced
to an Endpoint directly by AntreaProxy.

The default behavior is unchanged, i.e. proxyLoadBalancerIPs is true by
default.

Note that setting ProxyLoadBalancerIPs to false usually only makes sense
when ProxyAll is set to true and kube-proxy is removed from the cluser,
otherwise kube-proxy will still load-balance this traffic.

This feature cannot be leveraged by an external Load Balancer that
depends on the local proxy (AntreaProxy) being able to load balance
traffic destined to the External IP. This is the case for example with
MetalLB. However Load Balancers that rely on NodePort / NodePortLocal
can leverage this feature.

Fixes #3121

Signed-off-by: Antonin Bas <abas@vmware.com>