Remove all legacy (*.antrea.tanzu.vmware.com) APIs
These APIs were scheduled for deletion in December 2021. So it seems
reasonable to remove them in Antrea v1.6, without causing significant
disruption to users.

As part of this, the CRD mirroring controller code can be removed
entirely and the legacyCRDMirroring config option for the controller is
deprecated (it's a no-op and users trying to set this option to anything
will see a warning in the logs).

The API dcoumentation is updated. We add a note to inform users who may
still be using Antrea pre-v1.0 that they will need to make an
intermediate upgrade first and migrate their CRDs, if they want to
upgrade to Antrea >= v1.6.

Fixes #3298

Signed-off-by: Antonin Bas <abas@vmware.com>