Fix FlowAggregator e2e test random failure
In this commit, we fixed some random failure cases of FlowAggregator
e2e test.
One fix is for ToExternalFlows test. We give it 2 more seconds for
waiting flow records received on IPFIX collector.
Another fix is adding flowEndReason==3 in the condition of receiving
the last flow record.
A third fix is setting a bandwidth target of iperf command to make the
throughput test more stable.
More explanation could be found in the description of PR.

Signed-off-by: Yongming Ding <dyongming@vmware.com>