[Windows] Delete virtual network adapter before HNSEndpoint
Since it is possible that HNSEndpoint deletion is stuck because of
Windows HNS call issues, the virtual network adapter is disconnected
from the VMSwitch after OVS port is deletion. And then the
disconnected adapter is failed to remove forever.

This patch is to change the order to remove the virtual network adapter
first in CmdDel request to avoid the stale adapters.

Signed-off-by: wenyingd <wenyingd@vmware.com>