Fix NodePort e2e test (#3347)
In NodePort e2e test, a test Service and its backend Pod are
created. The match label is app=agnhost. However, the test
client is also a Pod that has label app=agnhost, but netexec is
not running on the test client. As a result, the test has a 50%
chance of failing. This PR fixes the issue by using a busybox Pod
as test client.

Add retry to wget in some tests.

Signed-off-by: Hongliang Liu <lhongliang@vmware.com>