Periodic Syncing of NPL iptables Rules
The iptables rules can be deleted by mistake. This patch is used to
call a function to ensure that the rules are restored automatically
in such a situation.

Also integration test was added.It adds some NPL iptables rules, and delete the
rules and check if the rules are recovered automatcally.

fixes antrea-io#2210