Remove unused job in Benchmark Github Workflow
The "check-changes" job is both incorrect and unused.
Incorrect because it ignores changes in the pkg/ directory, which is the
opposite of what it should be doing. This is probably because of poor
naming of the "args" action parameter for the has-changes action (it
should really be something like "ignore-paths").
Unused because the "go-benchmark-checks" will anyway run for all "push"
changes, and the Github workflow itself is configured to only run for
"push" events (on the main branch).

Signed-off-by: Antonin Bas <abas@vmware.com>