Fix issue of TestReconcileGatewayRoutesOnStartup in particular case
When a Service NodePort and an Egress CRD has the same backend Pod, accessing
to the NodePort Service may fail in particular cases. Assume that the backend
When proxyAll is enabled and a LoadBalancer Service that has external IPs is
created, routing entries like `192.168.1.1 via 169.254.0.253 dev antrea-gw0`
will be installed on K8s Node. In this situation, running
TestReconcileGatewayRoutesOnStartup will get a failure since the destination
is an IP without network mask and cannot be parsed to a CIDR.

Signed-off-by: Hongliang Liu <lhongliang@vmware.com>