Traffic to in-cluster destination should skip EgressMark table
Only traffic to external traffic is supposed to go to EgressMark table
as the table either redirects traffic to switching stage or drops it,
skipping postRouting stage, in which phase SNAT is performed if
required.

Without this patch, traffic to remote Pods via gateway would go to
EgressMark table when Egress feature is enabled. If the traffic needs
to be SNATed, e.g. when an external client accesses a NodePort service
which has an endpoint on a Node different from the accessed Node, there
would be a connnection issue.

Signed-off-by: Quan Tian <qtian@vmware.com>