Add function priorityAdd in pkg/agent/openflow/pipeline.go
Function priorityAdd is used to get a priority by adding an offset to a
defined priority. Note that, for flows of non-NetworkPolicy, only when all
existing priorities (priorityHigh, priorityNormal, priorityLow and
priorityMiss) cannot meet the design requirement and a new priority has to
be introduced, then priorityAdd should be used. Adding offset directly to
a defined priority is not recommended when building a flow because it is not
easy to track the priorities like `priorityHigh + 1`, `priorityNormal + 2`
for developers.

Signed-off-by: Hongliang Liu <lhongliang@vmware.com>