Fix race condition in ConntrackConnectionStore and FlowExporter
Conntrack connection store's polling go routine and flow exporter both access
to conntrack connection store, and there's a race condition error.

In the polling go routine, `deleteIfStaleOrResetConn` and `AddOrUpdateConn`
both grab the lock, modify `conn.IsPresent` field, and release the lock. Between
the execution of these two functions, it is likely that FlowExporter's timer is
triggered and it reads the wrong `conn.IsPresent` value in an intermidiate state.
We fix it by holding the lock until we finish the execution of both two functions.

Fixes: #3650

Signed-off-by: heanlan <hanlan@vmware.com>