Improve robustness of memberlist cluster unit tests (#3658)
The current wait condition used to determine whether an external IP has
been processed is not accurate, and this can lead to failures on slower
machine. Instead, we wrap calls to `ShouldSelectIP` inside a
`wait.Poll`, and wait for the function to return the expected result,
with the test failing in case of a timeout. We keep using `wait.Poll`
(instead of `wait.PollImmediate`) with a pretty long initial sleep
(100ms), as it is useful for "negative tests", which expect
`ShouldSelectIP` to return an error.

Fixes #3657

Signed-off-by: Antonin Bas <abas@vmware.com>