Add AppliedTo struct for internal NetworkPolicy
Currently AppliedTo shares the NetworkPolicyPeer with To/From fields,
yet some members such as Namespaces and FQDN would not be set in AppliedTo,
and Service would not be set in To/From.

So, it is beneficial to use a separate struct for distinction,
as well as future extensions. Current solution introduces the
AppliedTo struct for appliedto fields, which does not share
any member with NetworkPolicyPeer.

Fixes #2686.

Signed-off-by: Qiyue Yao <yaoq@vmware.com>