Add Policy Recommendation Antctl CLI in Flow Aggregator
In this commit, we add the command line interfaces of policy
recommendation feature inside the Flow Aggregator with Antctl.
There are 3 commands we add for policy recommendation, including start,
check status and get result of recommendation job. To use these
commands, deployment of Theia is required. Please find the document in
Theia repo for introduction of these commands.

Signed-off-by: Yongming Ding <dyongming@vmware.com>